StateFlow, Compose State
Now using Compose State and `StateFlow` in `MainViewModel` and in composables in `MainActivity`

Got rid of network security config and instead set `usesCleartextTraffic="true"` FOR NOW. Reason: I haven't set up TLS for the compute engine VM yet, it doesn't don't have a static IP, and it is really tedious having to change the path all the time in paths.properties.

Added a log line to indicate sockets configurations when they are initialized.